chore: remove AsyncHooksContextManager
