+++ packages/opentelemetry-context-async-hooks/src/index.ts
@@ -3,5 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { AsyncHooksContextManager } from './AsyncHooksContextManager';
 export { AsyncLocalStorageContextManager } from './AsyncLocalStorageContextManager';
